Feature:upgrade version

+++ aggregation.database/src/test/java/com/alibaba/smart/framework/engine/test/process/Order.java
@@ -39,6 +39,6 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 public class Order {
 
 
-    private Long yzje;
+    private Double yzje;
 
 }
